requirements.txt
aiohttp
fastapi
llmapi_cli==1.0.7
pycryptodome
pydantic
uvicorn
